dev-requirements.txt
flake8
django-nose
rednose
